% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_freq.R
\name{barplot_freq}
\alias{barplot_freq}
\title{Bar Plot of Frequencies}
\usage{
barplot_freq(
  data,
  sort_by = c("freq", "name"),
  decreasing = TRUE,
  flip = TRUE,
  limit_freq = NULL,
  top = NULL,
  bar_color = "steelblue",
  show_n = FALSE,
  label_color = "white",
  ...
)
}
\arguments{
\item{data}{A data frame or tibble containing the frequency data of the
taxonomic rank}

\item{sort_by}{Character vector specifying the sorting criteria for the bars.
Default is "freq". Valid options are "freq" and "name".}

\item{decreasing}{Logical indicating whether to sort the bars in decreasing
order. Default is TRUE.}

\item{flip}{Logical indicating whether to flip the orientation of the plot.
Default is TRUE.}

\item{limit_freq}{Numeric value specifying the minimum frequency threshold
for including a taxa in the plot. Default is NULL, which includes all taxa.}

\item{top}{Numeric value specifying the maximum number of taxa to include in
the plot. Default is NULL, which includes all taxa.}

\item{bar_color}{Color of the bars. Default is "steelblue".}

\item{show_n}{Logical indicating whether to display the records number over
the bar. Default is FALSE.}

\item{label_color}{Color of the labels. Default is "white".}

\item{...}{Additional arguments to be passed to \code{\link[=ggplot2]{ggplot2()}} and related functions.}
}
\value{
A ggplot object representing the bar plot of frequencies.
}
\description{
Generate a bar plot of frequencies for taxonomic ranks
}
