% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borreguiles.R
\docType{data}
\name{borreguiles}
\alias{borreguiles}
\title{Dataset: Phenology of flora of mediterranean high-mountains meadows (Sierra Nevada)}
\format{
A data frame with 11002 observations and 37 variables:
\itemize{
\item id: The ID of the observation.
\item modified: The modification date of the observation.
\item language: The language of the observation.
\item institutionCode, collectionCode: The institution and collection code.
\item datasetName: The name of the dataset.
\item basisOfRecord: The basis of the record.
\item occurrenceID: The occurrence ID.
\item catalogNumber: The catalog number.
\item recordedBy, recordedByID: The name and ID of the person who recorded the observation.
\item eventDate: The date of the observation.
\item year, month, day: The year, month and day of the observation.
\item continent: The continent where the observation was made.
\item country: The country where the observation was made.
\item countryCode: The country code.
\item stateProvince: The state or province where the observation was made.
\item county: The county where the observation was made.
\item locality: The locality of the observation.
\item minimumElevationInMeters, maximumElevationInMeters: The minimum and maximum elevation in meters.
\item decimalLatitude, decimalLongitude: The decimal latitude and longitude of the observation.
\item geodeticDatum: The geodetic datum.
\item coordinateUncertaintyInMeters: The coordinate uncertainty in meters.
\item scientificName: The scientific name of the plant species.
\item kingdom, phylum, class, order, family: The kingdom, phylum, class, order, family and genus of the taxa.
\item specificEpithet: The specific epithet of the plant species.
\item infraspecificEpithet: The infraspecific epithet of the plant species.
\item scientificNameAuthorship: The scientific name authorship.
}
}
\source{
Zamora R, Pérez-Luque AJ (2020). Dataset of Phenology of flora of
mediterranean high-mountains meadows (Sierra Nevada). Version 1.5.
Sierra Nevada Global Change Observatory. Andalusian Environmental Center,
University of Granada, Regional Government of Andalusia. Occurrence dataset
\href{https://doi.org/10.15468/qhqzub}{doi: 10.15468/qhqzub}

Pérez-Luque AJ, Sánchez-Rojas CP, Zamora R, Pérez-Pérez R, Bonet FJ (2015).
Dataset of Phenology of Mediterranean high-mountain meadows flora (Sierra Nevada,
Spain). PhytoKeys 46: 89-107. \href{https://doi.org/10.3897/phytokeys.46.9116}{doi: 10.3897/phytokeys.46.9116}
}
\usage{
borreguiles
}
\description{
This dataset contains information about the phenology of flora in Mediterranean
high-mountains meadows in the Sierra Nevada region.
}
\keyword{datasets}
