% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_cov.R
\name{cov_temporal}
\alias{cov_temporal}
\title{Calculate the temporal Coverage}
\usage{
cov_temporal(
  x,
  date_var,
  date_format = NULL,
  tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d"),
  report = TRUE
)
}
\arguments{
\item{x}{A data frame.}

\item{date_var}{The name of the date variable column in the data frame.}

\item{date_format}{The format of the date variable (optional).}

\item{tryFormats}{A vector of possible date formats to try if the date variable is not in a standard format.}

\item{report}{Logical value indicating whether to print the temporal coverage report.}
}
\value{
A data frame with the minimum and maximum dates from the date variable.
}
\description{
This function calculates the temporal coverage of a dataset (occurrences, events, or similar) based on a specified date variable.
}
\details{
If the \code{date_format} argument is not specified, the function will attempt to determine the format
of the date variable automatically using the \code{tryFormats} vector. If the date variable is not in a standard
format, an error will be thrown.
}
\examples{

data("borreguiles")
cov_temporal(borreguiles, date_var = "eventDate", report = TRUE)

}
