% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_taxonomy.R
\name{report_taxonomy}
\alias{report_taxonomy}
\title{Creates a report for a taxa rank}
\usage{
report_taxonomy(x, top = 5)
}
\arguments{
\item{x}{A data frame or tibble containing the frequency information of the taxa rank.
It must have columns 'n' and 'freq'. This dataframe typically generated
from \code{\link[=taxonomic_cov]{taxonomic_cov()}}}

\item{top}{The number of top taxa to include in the report.
Default value is 5.}
}
\value{
A character string representing the generated report.
}
\description{
Generate a summary report of the taxonomy of a taxa rank
}
\details{
This function generates a report based on the input data frame 'x'
containing frequency information of certain taxa rank. It provides
information about the most represented taxa.
}
\examples{
# Example usage
data(borreguiles)
# Calculate taxonomic coverage for scientificName and genus
d <- taxonomic_cov(borreguiles, category = c("scientificName", "genus"))
# Generate report
report_taxonomy(d$genus)

}
