% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{taxo_reportApp}
\alias{taxo_reportApp}
\title{App to generate taxonomic report}
\usage{
taxo_reportApp()
}
\description{
App to generate taxonomic report
}
\examples{
\dontrun{
data(borreguiles)
taxo_reportApp()
}

}
