% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_cov.R
\name{taxonomic_cov}
\alias{taxonomic_cov}
\title{taxonomic_cov}
\usage{
taxonomic_cov(
  x,
  category = c("scientificName", "kingdom", "phylum", "class", "order", "family",
    "subfamily", "genus", "all"),
  code_unknown = NA
)
}
\arguments{
\item{x}{A dataframe}

\item{category}{a character vector with taxonomic ranks defined by
\href{https://dwc.tdwg.org/terms/}{DarwinCore} standard.
#' @param code_unknown a custom string that identified the unclassified records
at taxa level (default to NA). If the dataset contains a custom string to
identify the unclassified taxa, please specify it here, \emph{e.g.} "unclass.",
"unknown". See \code{\link[=taxonomic_freq]{taxonomic_freq()}}}
}
\value{
A named list of tibbles summarizing the taxonomic coverage for each category.
}
\description{
It generates the taxonomic coverage of the dataset
}
\details{
It generates the taxonomic coverage of a occurrence dataset. For each
category, the record numbers and relative frequencies of each taxa are computed. The
category argument could be one of: \code{scientificName}, \code{kingdom}, \code{phylum}, \code{class},
\code{order}, \code{family}, \code{subfamily} and \code{genus}. For get coverage in all categories
type \code{all}.
}
\examples{
# Example usage
data(borreguiles)
# Calculate taxonomic coverage for scientificName and genus
taxonomic_cov(borreguiles, category = c("scientificName", "genus"))

}
