% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_freq.R
\name{taxonomic_freq}
\alias{taxonomic_freq}
\title{taxonomic_freq}
\usage{
taxonomic_freq(x, category, code_unknown = NA)
}
\arguments{
\item{x}{A dataframe}

\item{category}{a character vector with taxonomic ranks defined by
\href{https://dwc.tdwg.org/terms/}{DarwinCore} standard. One of \code{kingdom},
\code{phylum}, \code{class}, \code{order}, \code{family}, \code{subfamily}, \code{genus}, \code{scientificName}}

\item{code_unknown}{a custom string that identified the unclassified records
at taxa level (default to NA). If the dataset contains a custom string to
identify the unclassified taxa, please specify it here, \emph{e.g.} "unclass.",
"unknown"}
}
\value{
A tibble with the frequency of records (\code{n}) by taxonomic category,
and:
\itemize{
\item \code{freq_all}: percentage of records computed considering all the records
(\emph{i.e.} inlcuding the unknown or unclassified taxa)
\item \code{freq}: percentage of records computed excluding the unknown or unclassified
records
}
}
\description{
Compute the count and frequence of a taxonomic category
}
