% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifexplorer-package.R
\docType{package}
\name{gbifexplorer-package}
\alias{gbifexplorer}
\alias{gbifexplorer-package}
\title{gbifexplorer: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajpelu/gbifexplorer}
  \item Report bugs at \url{https://github.com/ajpelu/gbifexplorer/issues}
}

}
\keyword{internal}
