% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_taxo.R
\name{taxonomic_cov}
\alias{taxonomic_cov}
\title{taxonomic_cov}
\usage{
taxonomic_cov(
  x,
  category = c("scientificName", "kingdom", "phylum", "class", "order", "family",
    "subfamily", "genus", "all")
)
}
\arguments{
\item{x}{A dataframe}

\item{category}{taxonomic ranks defined by DarwinCore standard}
}
\value{
a list of ...
}
\description{
taxonomic_cov
}
